+++ packages/cli/test/fixtures/rules/a11y/dialog-nuxt-fail/app/pages/index.vue
@@ -1,0 +1,7 @@
+<template>
+  <Dialog>
+    <DialogContent>
+      <p>Auto-imported, no title.</p>
+    </DialogContent>
+  </Dialog>
+</template>